+++ addons/yat/src/commands/builtin/Man.cs
@@ -1,3 +1,4 @@
+using System;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Helpers;
@@ -15,11 +16,13 @@ namespace YAT.Commands
 		{
 			var commandName = data.RawData[1];
 
-			if (!data.Yat.CommandManager.Commands.TryGetValue(commandName, out var command))
+			if (!data.Yat.CommandManager.Commands.TryGetValue(commandName, out Type value))
 			{
 				data.Terminal.Output.Error(Messages.UnknownCommand(commandName));
 				return CommandResult.InvalidCommand;
 			}
+
+			ICommand command = (ICommand)Activator.CreateInstance(value);
 
 			// Check if the command manual is already in the cache.
 			if (cache.Get(commandName) is string manual)